delete CRA
