+++ src/App.css
@@ -1,39 +1,63 @@
-.App {
-  text-align: center;
+* {
+  padding: 0;
+  margin: 0;
+  box-sizing: border-box;
 }
 
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
-}
-
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
+body {
+  height: 100vh;
   display: flex;
   flex-direction: column;
-  align-items: center;
   justify-content: center;
-  font-size: calc(10px + 2vmin);
+  align-items: center;
+}
+
+.Application {
+  width: 45vw;
+}
+
+.answer {
+  display: flex;
+  align-items: center;
+  justify-content: flex-end;
+  height: 50px;
+  background-color: grey;
+  padding: 8px;
   color: white;
 }
 
-.App-link {
-  color: #61dafb;
+.operations {
+  display: grid;
+  grid-template-columns: auto auto auto auto;
+  grid-template-rows: repeat(5, 60px);
 }
 
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
+button {
+  cursor: pointer;
+  border: 1px solid #c7c7c7;
+  background-color: #e0e0e0;
+}
 
-  to {
-    transform: rotate(360deg);
-  }
+button:hover {
+  background-color: #e0e0e0b9;
+}
+
+button:active {
+  background-color: #80808088;
+}
+
+.zero {
+  grid-column: 1/span 2;
+}
+
+.maths {
+  background-color: #f5913e;
+}
+
+.maths:hover {
+  background-color: #ee832bc4;
+}
+
+.maths:active {
+  background-color: rgba(82, 40, 3, 0.966);
 }
