+++ src/App.js
@@ -14,19 +14,3 @@ class App extends React.PureComponent {
 }
 
 export default App;
-import React from 'react';
-import Calculator from './components/Calculator';
-import './App.css';
-
-class App extends React.PureComponent {
-  render() {
-    return (
-      <div className="Application">
-        <div className="answer">0</div>
-        <Calculator />
-      </div>
-    );
-  }
-}
-
-export default App;
