+++ src/App.test.js
@@ -1,8 +1,16 @@
-import { render, screen } from '@testing-library/react';
-import App from './App';
+import React from 'react';
+import Calculator from './components/Calculator';
+import './App.css';
 
-test('renders learn react link', () => {
-  render(<App />);
-  const linkElement = screen.getByText(/learn react/i);
-  expect(linkElement).toBeInTheDocument();
-});
+class App extends React.PureComponent {
+  render() {
+    return (
+      <div className="Application">
+        <div className="answer">0</div>
+        <Calculator />
+      </div>
+    );
+  }
+}
+
+export default App;
